Mint-X: Show a hint instead of a login box when no user is selected.
People were getting used to typing their password and pressing enter. When no user was selected, this would result in the password being shown on the screen as the name of the selected user.

It was important to retain the ability to enter a username which wasn't in the list, so this commit removes the entry by default but the user can get it back by pressing F1.

By default, the screen now shows a hint which invites you to select a user or to press F1 to enter a username.